+++ common/src/main/java/net/mehvahdjukaar/amendments/client/colors/SoftFluidColor.java
@@ -17,6 +17,9 @@ public class SoftFluidColor implements BlockColor {
                 if (tint == 1) {
                     var tank = bh.getSoftFluidTank();
                     if (tank.getFluid().is(BuiltInSoftFluids.MUSHROOM_STEW.get())) return -1;
+                    if (tank.getFluid().is(BuiltInSoftFluids.SUS_STEW.get())) return -1;
+                    if (tank.getFluid().is(BuiltInSoftFluids.RABBIT_STEW.get())) return -1;
+                    if (tank.getFluid().is(BuiltInSoftFluids.MUSHROOM_STEW.get())) return -1;
                     return tank.getTintColor(level, pos);
                 }
             }