+++ changelog.md
@@ -1,3 +1,4 @@
-- fixed jukebox model not being applied on first reload
-- waterlogged lilypad conversion will now happen seamlessly (hitbox and models dont appear to move)
-- waterlogging a carpeted block will make the carpet pop off
+- added tool hook. Shift click on a tripwire hook with a tool to place it there
+- added custom textures for soups and stew in cauldrons
+- improved wall lanterns models and renderers
+- fixed many minor issues